exercise 4.21*

+++ part_4/blog_list/controllers/blogs.js
@@ -60,8 +60,24 @@ blogsRouter.put('/:id', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const userId = decodedToken.id
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog === null) return response.status(404).json({error: "page not found or has already been deleted"})
+
+  if (blog.user.toString() === userId.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()  
+  }
+  else {
+    response.status(401).json({ error: 'invalid user' }).end()
+  }
 })
 
 module.exports = blogsRouter
